fix : 사용자 논리적삭제시 해당배송담당자도 같이, 논리삭제,기타코드수정
