+++ user/src/main/java/com/_hateam/user/presentation/controller/DeliverController.java
@@ -27,7 +27,6 @@ public class DeliverController {
     }
 
 
-
     // 검색(페이징추가, 권한별 분리)
     @GetMapping("/admin/search")
     public ResponseEntity<?> searchDeliverUsers(
